Better handling of synchronous helpers with resolved arguments. Remove q-deep.
In a completely synchronous setup, a block helper is now guaranteed to receive the real string
from `options.fn()` containing no promises. It can therefore modify the string (without
the risk of dealing with placeholders)

Helper arguments are not resolved with `q-deep` anymore, because this is to slow in large template
structures with large datasets. Instead, only each parameter (but no recursively) and each hash-arguments
is resolved.