Markers were not always resolved when block-helpers returned promise (#9)
Fixes #7

Consider a block-helper returning a promise, which is called
without promise parameters and which contained async-helpers
in its body.
In this case, the "prepareAndResolveMarkers" function would not
create a new "Markers" instance and would not resolve the
markers created in the "options.fn(this)" call. The marker of
the inner helper would remain unresolved.
The solution is to replace markers recursively. When a replacement
text is substituted for a marker, if is recursively searched for
markers as well.